saga_gui, postgres data source browser: remember login, sort tables by name (not by geometry type)

+++ saga-gis/src/saga_core/saga_gui/wksp_shapes.cpp
@@ -1165,7 +1165,7 @@ void CWKSP_Shapes::On_Draw(CWKSP_Map_DC &dc_Map, int Flags)
 			}
 		}
 
-		for(iShape=0; iShape<Get_Shapes()->Get_Selection_Count(); iShape++)
+		for(iShape=0; iShape<(int)Get_Shapes()->Get_Selection_Count(); iShape++)
 		{
 			if( iShape != m_Edit_Index )
 			{
